requirements.txt
# Core
python-dotenv
streamlit

# LangChain (Pydantic v2-compatible stack)
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-groq
langchain-objectbox

# OpenAI + Google

google-generativeai   # ✅ pinned to match langchain-google-genai

# Vectorstores + DBs
chromadb
faiss-cpu
cassio

# RAG helpers
bs4
beautifulsoup4
pypdf
sentence-transformers

# Tools
arxiv
wikipedia
langchain_exa
flask

# Other
groq
huggingface-hub